Resolved minor swift code issues in template files Added java tests for service and message generators
